+++ Common/DataModels/SchedulerRequest.cs
@@ -56,9 +56,9 @@
 
         public string RegistrationNumber { get; set; }
 
-        public int RequestedQuantity { get; set; }
+        public double RequestedQuantity { get; set; }
 
-        public int ReservedQuantity { get; set; }
+        public double ReservedQuantity { get; set; }
 
         public string Sub1 { get; set; }
