+++ WMS/Data/Core/Models/ItemListRow/ItemListRow.cs
@@ -17,7 +17,7 @@ namespace Ferretto.WMS.Data.Core.Models
 
         public DateTime CreationDate { get; set; }
 
-        public int DispatchedQuantity { get; set; }
+        public double DispatchedQuantity { get; set; }
 
         public string ItemDescription { get; set; }
 
@@ -29,7 +29,7 @@ namespace Ferretto.WMS.Data.Core.Models
 
         public int? Priority { get; set; }
 
-        public int RequestedQuantity { get; set; }
+        public double RequestedQuantity { get; set; }
 
         public ItemListRowStatus Status { get; set; }
